Avoid DerivingVia clashing with Safe on GHC 9.0.2+
As of GHC 9.0.2, `DerivingVia` is disallowed with `Safe` (see
https://gitlab.haskell.org/ghc/ghc/-/merge_requests/5675). As a result,
`hkd` will produce the following warning when compiled with GHC 9.0.2 or
higher:

```
on the commandline: warning:
    -XDerivingVia is not allowed in Safe Haskell; ignoring -XDerivingVia
```

The fix is simple: instead of enabling `DerivingVia` everywhere, only enable
it where it is needed. In fact, it turns out that `hkd` doesn't actually use
`DerivingVia` anywhere—it only needs `DerivingStrategies`.